LG LM-K310IM_MT6762_Android_10
LG LM-K310IM_MT6762_Android_10 is a smartphone model that runs on the Android 10 operating system. It has a 6.2-inch HD+ display, a 13-megapixel rear camera, and a 5-megapixel front camera. It is powered by a MediaTek MT6762 processor and has 3 GB of RAM and 32 GB of internal storage. It also supports dual SIM cards, microSD card expansion, and a fingerprint sensor.